src/main/java/baseball/domain/Number.java
package baseball.domain;

import java.util.Objects;

public class Number {
    private static final int MIN_NUMBER = 1;
    private static final int MAX_NUMBER = 9;

    private final int value;

    public Number(int value) {
        if (value < MIN_NUMBER || value > MAX_NUMBER) {
            throw new IllegalArgumentException("숫자는 1부터 9까지의 수만 허용됩니다.");
        }
        this.value = value;
    }

    public int getValue() {
        return value;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Number number = (Number) o;
        return getValue() == number.getValue();
    }

    @Override
    public int hashCode() {
        return Objects.hash(getValue());
    }
}
